Move event creation to action package
