+++ internal/action/action_finished_event_handler.go
@@ -50,13 +50,13 @@ func (eh *ActionFinishedEventHandler) HandleEvent(ctx context.Context) error {
 	// Additionally to the problem comment, send Info and Configuration Change Event to the entities in Dynatrace to indicate that remediation actions have been executed
 	if eh.event.GetStatus() == keptnv2.StatusSucceeded {
 
-		dtConfigEvent := dynatrace.CreateConfigurationEventDTO(eh.event, imageAndTag, eh.attachRules)
+		dtConfigEvent := createConfigurationEventDTO(eh.event, imageAndTag, eh.attachRules)
 		dtConfigEvent.Description = "Keptn Remediation Action Finished"
 		dtConfigEvent.Configuration = "successful"
 
 		dynatrace.NewEventsClient(eh.dtClient).AddConfigurationEvent(ctx, dtConfigEvent)
 	} else {
-		dtInfoEvent := dynatrace.CreateInfoEventDTO(eh.event, imageAndTag, eh.attachRules)
+		dtInfoEvent := createInfoEventDTO(eh.event, imageAndTag, eh.attachRules)
 		dtInfoEvent.Title = "Keptn Remediation Action Finished"
 		dtInfoEvent.Description = "error during execution"
