+++ internal/action/evaluation_finished_event_handler.go
@@ -33,7 +33,7 @@ func (eh *EvaluationFinishedEventHandler) HandleEvent(ctx context.Context) error
 
 	imageAndTag := eh.eClient.GetImageAndTag(eh.event)
 
-	ie := dynatrace.CreateInfoEventDTO(eh.event, imageAndTag, eh.attachRules)
+	ie := createInfoEventDTO(eh.event, imageAndTag, eh.attachRules)
 	qualityGateDescription := fmt.Sprintf("Quality Gate Result in stage %s: %s (%.2f/100)", eh.event.GetStage(), eh.event.GetResult(), eh.event.GetEvaluationScore())
 	ie.Title = fmt.Sprintf("Evaluation result: %s", eh.event.GetResult())
